\name{MixtureRegLTIC-package}
\alias{MixtureRegLTIC-package}
\docType{package}
\title{Mixture Regression Models for Left-Truncated and Interval-Censored Data}
\description{
Fit mixture regression models with nonsusceptibility/cure for left-truncated and interval-censored (LTIC) data (see Chen et al. (2013) <doi:10.1002/sim.5845>). This package also provides the nonparametric maximum likelihood estimator (NPMLE) for the survival/event curves with LTIC data.
}
\details{
\tabular{ll}{
Package: \tab MixtureRegLTIC\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2017-08-02\cr
Depends: \tab R (>= 3.3.1)\cr
License: \tab GPL-2|GPL-3\cr
Encoding: \tab UTF-8 \cr
LazyData: \tab yes\cr
}
}
\author{
Yuh-Chyuan Tsay, Chen-Hsin Chen, Ya-Chi Wu, Cheng-Fang Horng \cr
Maintainer: Yuh-Chyuan Tsay <yctsay@stat.sinica.edu.tw>
}
\references{
Chen CH, Tsay YC, Wu YC and Horng CF. Logistic-AFT location-scale mixture regression models with nonsusceptibility for left-truncated and general interval-censored data. Statistics in Medicine, 2013; 32:4285–4305.

Frydman H. A note on nonparametric estimation of the distribution function from interval-censored and truncated observations. Journal of the Royal Statistical Society, Series B, 1994; 56:71–74.

Moore RJ. Algorithm AS. 187: derivatives of the incomplete gamma integral. Applied Statistics - Journal of the Royal Statistical Society, Series C, 1982; 31:330–333.

Turnbull BW. The empirical distribution function with arbitrarily grouped, censored and truncated data. Journal of the Royal Statistical Society, Series B, 1976; 38:290–295.
}
